feat(render): Initial backend

+++ src/main/java/com/hokusai/interfaces/TextCommand.java
@@ -3,7 +3,7 @@ package com.hokusai.interfaces;
 public interface TextCommand extends Command {
   double x();
   double y();
-  double size();
+  int size();
   Color color();
   Padding padding();
   Boolean wrap();
